// src/pages/FeedPage/components/Feed/components/FeedActivity/components/LikesButton/components/Modal/LikesModal.type.ts
import { Dispatch, SetStateAction } from "react";
import { FeedActivityProps } from "../../../../FeedActivity.type";
import { idList } from "@/src/types";

export interface Like {
  avatar: string | null;
  firstName: string;
  lastName: string;
  id: number | string;
  likesList: idList;
}

export type LikesModalProps = {
  isOpen: boolean;
  setIsOpen: Dispatch<SetStateAction<boolean>>;
  likesData: Array<Like>;
  title?: string;
} & Pick<FeedActivityProps, "totalLikes">;